Move Bean to the right package
