+++ src/main/java/fr/esupportail/esupstage/controllers/jsf/package-info.java
@@ -1,0 +1,1 @@
+package fr.esupportail.esupstage.controllers.jsf;
